minor fix in flow of interceptor
